Fetch for game rooms

+++ Paparazzi/app/src/main/java/com/bryanho/paparazzi/requests/GetMessagesRequest.java
@@ -4,7 +4,7 @@ import com.bryanho.paparazzi.objects.Game;
 import com.google.gson.annotations.SerializedName;
 
 public class GetMessagesRequest {
-    @SerializedName("gameid") private int gameId;
+    @SerializedName("gameid") private long gameId;
 
     public GetMessagesRequest(Game game) {
         gameId = game.getGameId();
